improve tests with some helper functions
