+++ contracts/modules/andromeda-gatekeeper-spendlimit/src/contract.rs
@@ -9,8 +9,8 @@ use andromeda_modules::unified_asset::LegacyOwnerResponse;
 use cosmwasm_std::{ensure, Api};
 #[cfg(not(feature = "library"))]
 use cosmwasm_std::{
-    entry_point, to_binary, Addr, BankMsg, Binary, Coin, Deps, DepsMut, Env,
-    MessageInfo, Response, StdError, StdResult,
+    entry_point, to_binary, Addr, BankMsg, Binary, Coin, Deps, DepsMut, Env, MessageInfo, Response,
+    StdError, StdResult,
 };
 
 use crate::error::ContractError as CustomError;
