+++ contracts/modules/andromeda-user-account/src/lib.rs
@@ -1,6 +1,8 @@
 pub mod contract;
 pub mod error;
 #[cfg(test)]
+pub mod tests_helpers;
+#[cfg(test)]
 pub mod tests_integration;
 
 pub use crate::error::ContractError;
